Python 3 migration

+++ source/setup/setup.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 
 import os.path
 import shutil
@@ -14,13 +14,14 @@ def clean_old_scopus_cache_file():
 
     # Check the age of the exsiting files - scopus doesnt allow old citations
     # so dump the whole file if it is too old.
-    cached_scopus_files_list = os.listdir(config.cache_dir + '/raw/scopus/')
-    for this_file in cached_scopus_files_list:
-        if abs(datetime.datetime.now() - datetime.datetime.fromtimestamp(os.stat(config.cache_dir + '/raw/scopus/' + this_file).st_mtime)) > datetime.timedelta(days=config.scopus_citation_max_age_days):
-            file_path = config.cache_dir + '/raw/scopus/' + this_file
-            print('Deleting: ' + file_path)
-            logging.info('Deleting: ' + file_path)
-            os.remove(file_path)
+    if os.path.exists(config.cache_dir + '/raw/scopus'):
+        cached_scopus_files_list = os.listdir(config.cache_dir + '/raw/scopus/')
+        for this_file in cached_scopus_files_list:
+            if abs(datetime.datetime.now() - datetime.datetime.fromtimestamp(os.stat(config.cache_dir + '/raw/scopus/' + this_file).st_mtime)) > datetime.timedelta(days=config.scopus_citation_max_age_days):
+                file_path = config.cache_dir + '/raw/scopus/' + this_file
+                print('Deleting: ' + file_path)
+                logging.info('Deleting: ' + file_path)
+                os.remove(file_path)
 ################################################################################
